[PLU-238] TILES-COLLAB: Refactor queries and mutations (#574)
### Problem

All the graphql queries and mutation files live in the same folder, and
it's getting too large.

### Solution

Move tiles related files to its subfolder for better organization. 

### Testing

Test that all imports still work
