+++ packages/frontend/src/graphql/mutations/tiles/delete-rows.ts
