+++ packages/frontend/src/graphql/mutations/tiles/delete-table.ts
