+++ packages/frontend/src/graphql/queries/tiles/get-table.ts
