+++ packages/frontend/src/pages/Tile/index.tsx
@@ -3,8 +3,8 @@ import { ITableMetadata, ITableRow } from '@plumber/types'
 import { useParams } from 'react-router-dom'
 import { useQuery } from '@apollo/client'
 import { Center, Flex, Spinner } from '@chakra-ui/react'
-import { GET_ALL_ROWS } from 'graphql/queries/get-all-rows'
-import { GET_TABLE } from 'graphql/queries/get-table'
+import { GET_ALL_ROWS } from 'graphql/queries/tiles/get-all-rows'
+import { GET_TABLE } from 'graphql/queries/tiles/get-table'
 
 import Table from './components/Table'
 import TableBanner from './components/TableBanner'
